tests/wpt/tests/scroll-animations/css/view-timeline-axis-computed.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/scroll-animations-1/#view-timeline-axis">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<style>
  #outer { view-timeline-axis: block, inline; }
  #target { view-timeline-axis: y; }
</style>
<div id=outer>
  <div id=target></div>
</div>
<script>
test_computed_value('view-timeline-axis', 'initial', 'block');
test_computed_value('view-timeline-axis', 'inherit', 'block, inline');
test_computed_value('view-timeline-axis', 'unset', 'block');
test_computed_value('view-timeline-axis', 'revert', 'block');
test_computed_value('view-timeline-axis', 'block');
test_computed_value('view-timeline-axis', 'inline');
test_computed_value('view-timeline-axis', 'y');
test_computed_value('view-timeline-axis', 'x');
test_computed_value('view-timeline-axis', 'block, inline');
test_computed_value('view-timeline-axis', 'inline, block');
test_computed_value('view-timeline-axis', 'block, y, x, inline');
test_computed_value('view-timeline-axis', 'inline, inline, inline, inline');

test(() => {
  let style = getComputedStyle(document.getElementById('target'));
  assert_not_equals(Array.from(style).indexOf('view-timeline-axis'), -1);
}, 'The view-timeline-axis property shows up in CSSStyleDeclaration enumeration');

test(() => {
  let style = document.getElementById('target').style;
  assert_not_equals(style.cssText.indexOf('view-timeline-axis'), -1);
}, 'The view-timeline-axis property shows up in CSSStyleDeclaration.cssText');
</script>
